+++ base/exports.jl
@@ -855,7 +855,7 @@ export
     listenany,
     ltoh,
     mark,
-    bytesavailable,
+    nbytesavailable,
     ntoh,
     open,
     pipeline,