+++ NEWS.md
@@ -690,7 +690,7 @@ Deprecated or removed
 
   * `read(::IO, ::Ref)` is now a method of `read!`, since it mutates its `Ref` argument ([#21592]).
 
-  * `nb_available` is now `bytesavailable` ([#25634]).
+  * `nb_available` is now `nbytesavailable` ([#25634]).
 
   * `skipchars(io::IO, predicate; linecomment=nothing)` is deprecated in favor of
     `skipchars(predicate, io::IO; linecomment=nothing)` ([#25667]).